# docs/swagger.yaml
definitions:
...
  controller.brandInput:
    properties:
      description:
        type: string
      logo_url:
        type: string
      name:
        type: string
    required:
    - logo_url
    - name
    type: object
...
  controller.commentInput:
    properties:
      content:
        type: string
    required:
    - content
    type: object
...
  models.Brand:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      logo_url:
        type: string
      name:
        type: string
      phones:
        items:
          $ref: '#/definitions/models.Phone'
        type: array
      updated_at:
        type: string
    type: object
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      review_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Phone:
    properties:
      brand_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      model:
        type: string
      price:
        type: integer
      release_date:
        type: string
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      specification:
        items:
          $ref: '#/definitions/models.Specification'
        type: array
      updated_at:
        type: string
    type: object
...
  models.Review:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      phone_id:
        type: integer
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Specification:
    properties:
      additional_feature:
        type: string
      battery:
        type: string
      camera:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      memory:
        type: integer
      network:
        type: string
      operating_system:
        type: string
      phone_id:
        type: integer
      storage:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /brands:
    get:
      description: Get a list of Phone brands.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Brand'
            type: array
      summary: Get all Phones brands. (PUBLIC)
      tags:
      - Brands
    post:
      description: Creating a new Brand data, only account with role admin can access
        this route
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: example JSON body to create a new Brand
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.brandInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Brand'
      security:
      - BearerToken: []
      summary: Create New Brand (ADMIN ONLY)
      tags:
      - Brands
  /brands/{id}:
    delete:
      description: Delete a Brand by id
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Brand id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete Brand by id . (ADMIN ONLY)
      tags:
      - Brands
    get:
      description: Get Brand data by ID.
      parameters:
      - description: Brand id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Brand'
            type: array
      summary: Get single brand data by ID. (PUBLIC)
      tags:
      - Brands
    put:
      description: Update Brand data by id, only account with role admin can access
        this route
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Brand id
        in: path
        name: id
        required: true
        type: string
      - description: tExample JSON body to update Brand data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.brandInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Brand'
      security:
      - BearerToken: []
      summary: Update Brand data. (ADMIN ONLY)
      tags:
      - Brands
  /brands/{id}/phones:
    get:
      description: Get all Phones data by brand id.
      parameters:
      - description: Brand id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Brand'
            type: array
      summary: Get phones data by Brand id. (PUBLIC)
      tags:
      - Brands
  /comments/{id}:
    delete:
      description: Delete a Comment by id, (user ID is taken from the JWT token)
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete Comment by id .
      tags:
      - Comments
...
  /reviews/{id}/comments:
    get:
      description: Get all Comments data by review id.
      parameters:
      - description: Review id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
      summary: Get comments data by Review id.
      tags:
      - Reviews
    post:
      description: Creating a comment on a phone review, (user ID is taken from the
        JWT token)
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: review id
        in: path
        name: id
        required: true
        type: string
      - description: example JSON body to create a comment, user_id is taken from
          the authorization token
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.commentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
      security:
      - BearerToken: []
      summary: Create Comment on a review
      tags:
      - Reviews
  /reviews/{id}/comments/{com_id}:
    put:
      description: This route will update comment data, will only be able to update
        data related to the logged in user (user ID is taken from the JWT token)
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: review id
        in: path
        name: id
        required: true
        type: string
      - description: comment id
        in: path
        name: com_id
        required: true
        type: string
      - description: example JSON body to update a comment, user_id is taken from
          the authorization token
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.commentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
      security:
      - BearerToken: []
      summary: Update Comment on a review
      tags:
      - Reviews
...
swagger: "2.0"